Introduce Vertx3RxGeneratorBase as basis for RX1 and RX2 generators that will differ from RX3
